加入聊天功能等等

+++ business/goods-service/src/main/java/com/study/shop/business/controller/GoodsController.java
@@ -19,6 +19,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.dubbo.config.annotation.Reference;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -33,10 +36,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Date;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * @author Tiger
@@ -50,25 +57,94 @@ public class GoodsController {
     private TbItemService tbItemService;
     @Reference(version = "1.0.0")
     private TbUserService tbUserService;
+    @Autowired
+    RedisTemplate redisTemplate;
 
+    private String ACCESS_LOG_KEY = "user_logs_";
+
+    /**
+     * 搜索推荐接口
+     *
+     * @param keyword 关键字
+     * @return 推荐列表
+     */
     @GetMapping("/recommend/{keyword}")
     public ResponseResult<List<String>> getRecommend(@PathVariable String keyword) {
         return new ResponseResult<>(ResponseResult.CodeStatus.OK, tbItemService.searchRecommend(keyword));
     }
 
+    /**
+     * 搜索商品
+     *
+     * @param goodsSearchDTO 条件实体
+     * @return 搜索结果
+     */
     @GetMapping("/search")
     public ResponseResult<PageInfo<GoodsVO>> searchGoods(GoodsSearchDTO goodsSearchDTO) {
         return new ResponseResult<>(ResponseResult.CodeStatus.OK, tbItemService.searchItem(goodsSearchDTO));
     }
 
+    /**
+     * 获取商品详情
+     *
+     * @param goodsId goodsId
+     * @return 详情实体
+     */
     @GetMapping("/{goodsId}")
-    public ResponseResult<GoodDetailVO> goodsDetail(@PathVariable("goodsId") String goodsId) {
+    public ResponseResult<Object> goodsDetail(@PathVariable("goodsId") String goodsId) {
+        ZSetOperations zSetOperations = redisTemplate.opsForZSet();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        if (!"anonymousUser".equals(username)) {
+            //        在redis中保存用户足迹
+            zSetOperations.add(ACCESS_LOG_KEY + username, goodsId, System.currentTimeMillis());
+        }
         if (!StringUtils.isNumeric(goodsId)) {
             throw new BusinessException(ExceptionStatus.ILLEGAL_REQUEST);
         }
-        return new ResponseResult<>(ResponseResult.CodeStatus.OK, "获取商品详情", tbItemService.getGoodDetail(Long.valueOf(goodsId)));
+        GoodDetailVO goodDetail = tbItemService.getGoodDetail(Long.valueOf(goodsId));
+        return new ResponseResult<>(ResponseResult.CodeStatus.OK, "获取商品详情", goodDetail);
     }
 
+    @GetMapping("/logs")
+    public ResponseResult<Object> getMyLogs(@RequestParam(name = "page", defaultValue = "1") Integer page,
+                                            @RequestParam(name = "size", defaultValue = "8") Integer size) {
+        ZSetOperations<String, String> zSetOperations = redisTemplate.opsForZSet();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String username = authentication.getName();
+        int totalPage = 0;
+        int totalCount = 0;
+        List<GoodDetailVO> logsDetail = new ArrayList<>();
+        Set<String> logSet = new HashSet<>();
+        if (!"anonymousUser".equals(username)) {
+            //        在redis中获取用户足迹   精度丢失问题，先转为double
+            Double redisCount = zSetOperations.size(ACCESS_LOG_KEY + username).doubleValue();
+            double i = redisCount / size;
+            double redisPage = Math.ceil(i);
+            totalCount = redisCount.intValue();
+            totalPage = (int) redisPage;
+            logSet = zSetOperations.reverseRange(ACCESS_LOG_KEY + username, (page - 1) * size, page * size - 1);
+        }
+        if (logSet != null) {
+            List<Long> longList = new ArrayList<>();
+            for (String s : logSet) {
+                longList.add(Long.valueOf(s));
+            }
+            logsDetail = tbItemService.getLogsGoodsDetail(longList);
+        }
+        Map<String, Object> map = new HashMap<>();
+        map.put("logs", logsDetail);
+        map.put("totalCount", totalCount);
+        map.put("totalPage", totalPage);
+        return new ResponseResult<>(ResponseResult.CodeStatus.OK, "获取商品详情", map);
+    }
+
+    /**
+     * 获取我发布的商品
+     *
+     * @param myGoodsDTO myGoods
+     * @return 实体
+     */
     @PreAuthorize("hasAuthority('USER')")
     @GetMapping()
     public ResponseResult getMyGoods(MyGoodsDTO myGoodsDTO) {
@@ -91,7 +167,7 @@ public class GoodsController {
             throw new BusinessException(ExceptionStatus.ACCOUNT_NOT_EXIST);
         }
         GoodDetailVO myGoodsDetail = tbItemService.getMyGoodsDetail(tbUser.getId(), Long.valueOf(goodsId));
-        Map<String, Object> target = new HashMap<>(7);
+        Map<String, Object> target = new HashMap<>(7, 1);
         target.put("title", myGoodsDetail.getTitle());
         target.put("sellPoint", myGoodsDetail.getSellPoint());
         target.put("price", myGoodsDetail.getPrice());
@@ -113,6 +189,10 @@ public class GoodsController {
             throw new BusinessException(ExceptionStatus.ACCOUNT_NOT_EXIST);
         }
         Long targetId = Long.valueOf(goodsId);
+        GoodDetailVO goodDetail = tbItemService.getGoodDetail(Long.valueOf(goodsId));
+        if (goodDetail.getStatus() == 2 || goodDetail.getStatus() == 3) {
+            return new ResponseResult<>(ResponseResult.CodeStatus.FAIL, "已售出或被拍下商品不可删除！");
+        }
         int i = tbItemService.deleteGoods(tbUser.getId(), targetId);
         if (i > 0) {
             return new ResponseResult<>(ResponseResult.CodeStatus.OK);
@@ -155,6 +235,19 @@ public class GoodsController {
             return new ResponseResult<>(ResponseResult.CodeStatus.FAIL);
         }
         return new ResponseResult<>(ResponseResult.CodeStatus.OK);
+    }
+
+    @GetMapping("/other/{id}")
+    public ResponseResult getOtherGoods(@PathVariable(value = "id", required = true) String id) {
+        TbUser byId = tbUserService.getById(Long.valueOf(id));
+        Date date = new Date();
+        int days = (int) ((date.getTime() - byId.getCreateTime().getTime()) / (1000 * 3600 * 24));
+        Map<String, Object> map = new HashMap<>(6, 1);
+        map.putAll(tbItemService.getOtherGoodsNumberAndSellCount(Long.valueOf(id)));
+        map.put("nickName", byId.getNickName());
+        map.put("icon", byId.getIcon());
+        map.put("comeTime", days);
+        return new ResponseResult<>(ResponseResult.CodeStatus.OK, map);
     }
 
     @NotNull
